Moved logger to steps, added ProjectStep
